refactor: rename exam folder to subject folder and delete exam folder
