+++ src/app/(app)/subject/[slug]/session/page.tsx
@@ -1,4 +1,4 @@
-import SessionTable from "@/app/(app)/(exam)/_components/session-table";
+import SessionTable from "@/app/(app)/subject/_components/session-table";
 import { createClient } from "@/lib/supabase/client";
 
 export default async function SessionPage({
